% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateBarcodes.R
\name{associateBarcodes}
\alias{associateBarcodes}
\title{Associate barcode(s) to variants}
\usage{
associateBarcodes(
  countTable,
  barcodeCol,
  variantCol,
  countCol = "nbrReads",
  minCount = 10,
  minFracOfBarcodeCount = 0.5
)
}
\arguments{
\item{countTable}{A data frame containing barcode sequences, variant 
sequences and corresponding read counts for each barcode-variant pair.}

\item{barcodeCol, variantCol, countCol}{Character scalars indicating the 
names of the columns of \code{countTable} containing barcode 
sequences, variant IDs and read counts, respectively.}

\item{minCount}{Numeric scalar. Pairs with fewer reads will be flagged 
for exclusion.}

\item{minFracOfBarcodeCount}{Numeric scalar. Pairs where the read count 
is lower than \code{minFracOfBarcodeCount} * total barcode count will 
be flagged for exclusion.}
}
\value{
A list with two elements: \code{counts} (a data frame with counts)
and \code{plots} (a list of ggplot objects).
}
\description{
Associate barcode(s) to variants
}
\author{
Charlotte Soneson
}
