% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcErrorStats.R
\name{calcErrorStats}
\alias{calcErrorStats}
\title{Get error statistics from constant sequences}
\usage{
calcErrorStats(L, expectedForward = NULL, expectedReverse = NULL,
  verbose = FALSE)
}
\arguments{
\item{L}{List, as output by \code{readTransFastqs}.}

\item{expectedForward, expectedRevers}{Character(1) with the expected sequences
of the forward or reverse constant sequences. If \code{NULL}, the most
frequent sequence will be used.}

\item{verbose}{Logical(1), whether to print out progress messages.}
}
\value{
A list with 6 elements:
    \describe{
    \item{propErrorsConstantF}{proportion of mismatching bases in forward constant sequence}
    \item{propErrorsConstantR}{proportion of mismatching bases in reverse constant sequence}
    \item{QcountConstantErrorsF}{numbers of mismatching bases per quality score in forward constant sequence}
    \item{QcountConstantErrorsR}{numbers of mismatching bases per quality score in reverse constant sequence}
    \item{QcountConstantCorrectF}{numbers of matching bases per quality score in forward constant sequence}
    \item{QcountConstantCorrectR}{numbers of matching bases per quality score in reverse constant sequence} }
}
\description{
From the constant sequence parts, calculate sequencing error statistics by
comparison to expected sequences.
}
\author{
Michael Stadler
}
