% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestFastqs.R
\name{checkNumericInput}
\alias{checkNumericInput}
\title{Check numeric input argument}
\usage{
checkNumericInput(..., nonnegative)
}
\description{
Check that input argument is numeric, length 1 and non-negative.
}
\keyword{internal}
