% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseMutantsByAA.R
\name{collapseMutantsByAA}
\alias{collapseMutantsByAA}
\title{Collapse counts by mutated amino acid}
\usage{
collapseMutantsByAA(se, nameCol = "mutantNameAA")
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}
generated by \code{\link{summarizeExperiment}}}

\item{nameCol}{A character scalar providing the column of 
\code{rowData(se)} that contains the amino acid mutant names (that will 
be the new row names).}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} where
    counts have been aggregated by the mutated amino acid(s).
}
\description{
Collapse counts by mutated amino acid
}
\author{
Charlotte Soneson, Michael Stadler
}
