% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseSeqs.R
\name{collapseSeqs}
\alias{collapseSeqs}
\title{Collapse similar sequences}
\usage{
collapseSeqs(
  seqs,
  counts = NULL,
  method = c("greedy_clustering"),
  tol = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{seqs}{A \code{character} vector with sequences to be grouped. The
sequences have to be of equal length.}

\item{counts}{[optional] A \code{numeric} vector with observed counts for
each sequence in \code{seqs}. If provided, \code{seqs} will be sorted
decreasingly by \code{counts} (start collapsoing from the most frequent
sequence). If \code{NULL}, the sequences are collapsed in the order given.}

\item{method}{A \code{character} scalar selecting the collapsing method.
One of: greedy_clustering.}

\item{tol}{A \code{numeric} scalar defining the tolerance for similar
sequences. If \code{tol} is in [0, 1), it is interpreted as the fraction
of allowed mismatch bases (\code{ceiling(tol * nchar(seqs[1]))}). A value
greater or equal to 1 is rounded and interpreted as the maximum allowed
Hamming distance.}

\item{verbose}{A \code{logical} scalar. If \code{TRUE}, report on progress on
the R console.}
}
\value{
A \code{list} of \code{character} vectors. List elements are the
  grouped sequences, list names are the query sequences that defined the
  group.
}
\description{
Group sequences of equal length into sets of similar sequences,
  based on Hamming distance. An example application is the collapsing of
  unique molecular identifier variants that arise due to sequencing errors.
}
\details{
To speed-up the search for similar sequences, the function
  stores the sequences in a BK tree (https://en.wikipedia.org/wiki/BK-tree),
  which reduced the number of string distance calculations. For a general
  discussion of this problem with further details on BK trees see
  Liu, D. (2019).
  
  The following algorithms can be selected using the \code{method} argument:
  \itemize{
    \item greedy_clustering: staring from the first (most frequent)
      sequence, group it with all similar sequences that are within
      Hamming distance \code{tol}. Remove these sequences from \code{seqs}
      and continue with the next frequent sequence, until \code{seqs} is
      empty.
  }
}
\references{
Daniel Liu. "Algorithms for efficiently collapsing reads with
  Unique Molecular Identifiers". PeerJ. 2019; 7: e8275.
  doi: 10.7717/peerj.8275
  PMID: 31871845
  https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6921982/
}
\author{
Michael Stadler
}
