% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterReads.R
\name{filterReads}
\alias{filterReads}
\title{Filter FASTQ files from CIS or TRANS experiments}
\usage{
filterReads(L, avePhredMin = 20, variableNMax = 0, umiNMax = 0,
  wildTypeForward = NULL, wildTypeReverse = NULL,
  nbrMutatedCodonsMax = 1, forbiddenMutatedCodon = NULL)
}
\arguments{
\item{L}{SummarizedExperiment, as output by \code{readFastqs}.}

\item{avePhredMin}{numeric(1) Minimum average Phred score in the variable
region for a read to be retained. If L contains both forward and reverse
variable regions, the minimum average Phred score has to be achieved in
both for a read pair to be retained.}

\item{variableNMax}{numeric(1) Maximum number of Ns in the variable region
for a read to be retained.}

\item{umiNMax}{numeric(1) Maximum number of Ns in the UMI for a read to be
retained.}

\item{wildTypeForward, wildTypeReverse}{character(1), the wild type sequence
for the forward and reverse variable region.}

\item{nbrMutatedCodonsMax}{numeric(1) Maximum number of mutated codons that
are allowed.}

\item{forbiddenMutatedCodon}{character(1) A codon (can contain ambiguous
IUPAC characters, see \code{\link[Biostrings]{IUPAC_CODE_MAP}}). If a read
pair contains a mutated codon matching this pattern, it will be filtered
out.}
}
\value{
A filtered SummarizedExperiment object
}
\description{
Filter FASTQ files from CIS or TRANS experiments
}
\author{
Charlotte Soneson
}
