% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTrans.R
\name{filterTrans}
\alias{filterTrans}
\title{Filter FASTQ files from TRANS experiment}
\usage{
filterTrans(L, avePhredMin = 20, variableNMax = 0, umiNMax = 0,
  wildTypeForward = NULL, wildTypeReverse = NULL,
  nbrMutatedCodonsMax = 1, forbiddenMutatedCodon = NULL)
}
\arguments{
\item{L}{List, as output by \code{readTransFastqs}.}

\item{avePhredMin}{Numeric(1) Minimum average Phred score in the 
variable region for a read to be retained.}

\item{variableNMax}{Numeric(1) Maximum number of Ns in the 
variable region for a read to be retained.}

\item{umiNMax}{Numeric(1) Maximum number of Ns in the UMI for a 
read to be retained.}
}
\value{
A filtered list with eight elements:
\describe{
  \item{umis}{Merged forward and reverse UMI sequences}
  \item{constantSeqForward}{Constant forward sequence}
  \item{constantSeqReverse}{Constant reverse sequence}
  \item{variableSeqForward}{Variable forward sequence}
  \item{variableSeqReverse}{Variable reverse sequence}
  \item{minQualMutatedForward}{Minimal base quality of a mutated base in forward sequence}
  \item{minQualMutatedReverse}{Minimal base quality of a mutated base in reverse sequence}
  \item{readSummary}{data.frame tabulating the total number of read pairs, 
  and the number that match any of the filtering criteria.} 
}
}
\description{
Filter FASTQ files from TRANS experiment
}
\author{
Charlotte Soneson
}
