% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQCReport.R
\name{generateQCReport}
\alias{generateQCReport}
\title{Generate QC report}
\usage{
generateQCReport(se, outFile, forceOverwrite = FALSE)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object, typically generated with 
\code{summarizeExperiment()}.}

\item{outFile}{Character string providing the name of the output file. 
Should have the extension \code{.html}.}

\item{forceOverwrite}{Logical scalar, indicating whether an existing file
with the same name as \code{outFile} should be overwritten.}
}
\value{
Invisibly, the path to the generated html file.
}
\description{
Generate QC report
}
\author{
Charlotte
}
