% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{isValidL}
\alias{isValidL}
\title{Check if "sequences" argument is valid}
\usage{
isValidL(L)
}
\arguments{
\item{L}{A list as returned by \code{readFastqsTrans}}
}
\description{
Currently: check if it is a list with expected elements.
}
