% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReadPairs.R
\name{mergeReadPairs}
\alias{mergeReadPairs}
\title{Merge forward and reverse read pairs}
\usage{
mergeReadPairs(readsForward, readsReverse)
}
\arguments{
\item{read1, read2}{qualityScaledDNAStringSet objects containing matched
forward and reverse reads.}
}
\value{
A QualityScaledDNAStringSet with consensus sequences/qualities
}
\description{
Generate a consensus sequence for each pair of forward and reverse reads, for
each position choosing the base (forward/reverse) with the highest base
quality. If both reads have equal base quality, the forward read base is
selected. The function assumes that all reads (forward/reverse, and within
each object) have the same length.
}
\author{
Charlotte Soneson
}
