% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterReads.R
\name{mismatchFilter}
\alias{mismatchFilter}
\title{Filter reads based on mismatches with the wildtype sequence}
\usage{
mismatchFilter(wildTypeSeq, variableSeq, nbrMutatedCodonsMax = NULL,
  forbiddenMutatedCodon = NULL, maxChunkSize = 1e+05)
}
\arguments{
\item{wildTypeSeq}{character(1), the wild type sequence}

\item{variableSeq}{QualityScaledDNAStringSet, containing the sequences of the
variable region}

\item{nbrMutatedCodonsMax}{numeric(1) Maximum number of mutated codons that
are allowed.}

\item{forbiddenMutatedCodon}{character(1) A codon (can contain ambiguous
IUPAC characters, see \code{\link[Biostrings]{IUPAC_CODE_MAP}}). If a read
pair contains a mutated codon matching this pattern, it will be filtered
out.}

\item{maxChunkSize}{numeric(1), largest allowed chunk size for steps where
the reads are processed in chunks.}
}
\value{
A list with four elements:
  \describe{
    \item{qualMutated}{IntegerList with qualities for the mutated bases}
    \item{keepMM}{Logical vector, whether to keep each read}
    \item{basesWithMutations}{IntegerList, mutated base positions}
    \item{codonsWithMutations}{NumericList, mutated codons}
 }
}
\description{
Filter reads based on mismatches with the wildtype sequence
}
\author{
Charlotte Soneson
}
\keyword{internal}
