% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFastqs.R
\name{readFastqsInputCheck}
\alias{readFastqsInputCheck}
\title{Check inputs for readFastqs}
\usage{
readFastqsInputCheck(experimentType, fastqForward, fastqReverse,
  skipForward, skipReverse, umiLengthForward, umiLengthReverse,
  constantLengthForward, constantLengthReverse, variableLengthForward,
  variableLengthReverse, adapterForward, adapterReverse, verbose)
}
\arguments{
\item{experimentType}{character(1), either "cis" or "trans". If this is set
to "cis", the variable sequences from the forward and reverse reads will be
consolidated into one single sequence.}

\item{fastqForward}{character(1), paths to FASTQ files
corresponding to forward and reverse reads, respectively.}

\item{fastqReverse}{character(1), paths to FASTQ files
corresponding to forward and reverse reads, respectively.}

\item{skipForward}{numeric(1), the number of bases to skip in the
start of each forward and reverse read, respectively.}

\item{skipReverse}{numeric(1), the number of bases to skip in the
start of each forward and reverse read, respectively.}

\item{umiLengthForward}{numeric(1), the length of the
barcode (UMI) sequence in the forward/reverse reads, respectively, not
including the skipped bases (defined by
\code{skipForward}/\code{skipReverse}).}

\item{umiLengthReverse}{numeric(1), the length of the
barcode (UMI) sequence in the forward/reverse reads, respectively, not
including the skipped bases (defined by
\code{skipForward}/\code{skipReverse}).}

\item{constantLengthForward}{numeric(1), the length of
the constant sequence in the forward/reverse reads, respectively.}

\item{constantLengthReverse}{numeric(1), the length of
the constant sequence in the forward/reverse reads, respectively.}

\item{variableLengthForward}{numeric(1), the length of
the variable sequence in the forward/reverse reads, respectively.}

\item{variableLengthReverse}{numeric(1), the length of
the variable sequence in the forward/reverse reads, respectively.}

\item{adapterForward}{character(1), the adapter sequence for
forward/reverse reads, respectively. If a forward/reverse read contains the
corresponding adapter sequence, the sequence pair will be filtered out.
If set to \code{NULL}, no adapter filtering is performed. The number of
filtered read pairs are reported in the return value.}

\item{adapterReverse}{character(1), the adapter sequence for
forward/reverse reads, respectively. If a forward/reverse read contains the
corresponding adapter sequence, the sequence pair will be filtered out.
If set to \code{NULL}, no adapter filtering is performed. The number of
filtered read pairs are reported in the return value.}

\item{verbose}{logical(1), whether to print out progress messages.}
}
\description{
Check inputs for readFastqs
}
\keyword{internal}
