% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFastqsTrans.R
\name{readFastqsTrans}
\alias{readFastqsTrans}
\title{Read FASTQ files from TRANS experiment}
\usage{
readFastqsTrans(fastqForward, fastqReverse, skipForward = 1,
  skipReverse = 1, umiLengthForward = 10, umiLengthReverse = 8,
  constantLengthForward = 18, constantLengthReverse = 20,
  adapterForward = NULL, adapterReverse = NULL)
}
\arguments{
\item{fastqForward, fastqReverse}{FASTQ files corresponding to forward and
reverse reads, respectively.}

\item{skipForward, skipReverse}{Numeric(1), the number of bases to skip in the
start of each read.}

\item{umiLengthForward, umiLengthReverse}{Numeric(1), the length of the
barcode (UMI) sequence in the forward/reverse reads, respectively, not
including the skipped bases.}

\item{constantLengthForward, constantLengthReverse}{Numeric(1), the length of
the constant sequence in the forward/reverse reads, respectively.}

\item{adapterForward, adapterReverse}{Character(1), the adapter sequences, if
found in forward/reverse sequences, the sequence pair will be filtered out.
\code{NULL} does not perform any filtering. The number of filtered read
pairs are reported in the return value.}
}
\value{
A list with five elements: \describe{ \item{umis}{Merged forward and
  reverse UMI sequences} \item{constantSeqForward}{Constant forward sequence}
  \item{constantSeqReverse}{Constant reverse sequence}
  \item{variableSeqForward}{Variable forward sequence}
  \item{variableSeqReverse}{Variable reverse sequence}
  \item{numberReadPairsFiltered}{because of matches to adaptor sequence(s)} }
}
\description{
It is assumed that the forward and reverse reads are of the form UMI -
constant sequence - variable sequence, and that the length of the UMI
sequence, and the length of the constant sequence, is the same across all the
reads.
}
\author{
Charlotte Soneson
}
