% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{tabulateQualitiesByMatchstate}
\alias{tabulateQualitiesByMatchstate}
\title{Count numbers of matching/mismatching bases per quality score}
\usage{
tabulateQualitiesByMatchstate(pattern, subject, maxQ = 99L)
}
\arguments{
\item{pattern}{A nucleotide sequence corresponding to the 
expected sequence}

\item{subject}{A DNAStringSet with observed variable sequences}

\item{maxQ}{Maximum quality to tabulate (\code{nbins} argument to \code{tabulate})}
}
\value{
a two-element list with the numbers of matching/mismatching bases of a given quality
}
\description{
Based on the code from findMismatchpositions, tabulate base qualities
for matching/mismatching bases.
}
\keyword{internal}
