% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcNearestStringDist}
\alias{calcNearestStringDist}
\title{Calculate distances to the nearest string}
\usage{
calcNearestStringDist(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
An integer vector of the same length as \code{x}.
}
\description{
Given a character vector, calculate the distance for each element
to the nearest neighbor amongst all the other elements.
}
