% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRelativeFC.R
\name{calculateRelativeFC}
\alias{calculateRelativeFC}
\title{Calculate logFCs relative to WT using edgeR}
\usage{
calculateRelativeFC(se, design, coef, selAssay = "counts",
  pseudocount = 1)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{design}{Design matrix. The rows of the design matrix must be in the
same order as the columns in \code{se}.}

\item{coef}{Coefficient(s) to test with edgeR.}

\item{selAssay}{Assay to select from \code{se} for the analysis.}

\item{pseudocount}{Pseudocount to add when calculating log-fold changes.}
}
\description{
Calculate logFCs and associated p-values for a given comparison, using the
Negative Binomial quasi-likelihood framework provided by edgeR. The observed
counts for the WT variant will be used as offsets in the model.
}
\author{
Charlotte Soneson, Michael Stadler
}
