% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRelativeFC.R
\name{calculateRelativeFC}
\alias{calculateRelativeFC}
\title{Calculate logFCs relative to WT using edgeR}
\usage{
calculateRelativeFC(
  se,
  design,
  coef = NULL,
  contrast = NULL,
  WTrows,
  selAssay = "counts",
  pseudocount = 1,
  method = "edgeR",
  csawnorm = FALSE
)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{design}{Design matrix. The rows of the design matrix must be in the
same order as the columns in \code{se}.}

\item{coef}{Coefficient(s) to test with edgeR or limma.}

\item{contrast}{Numeric contrast to test with edgeR or limma.}

\item{WTrows}{Vector of row names that will be used as the reference when
calculating logFCs and statistics. If more than one value is provided, the
sum of the corresponding counts is used to generate offsets. If NULL,
offsets will be defined as the effective library sizes (using TMM
normalization factors).}

\item{selAssay}{Assay to select from \code{se} for the analysis.}

\item{pseudocount}{Pseudocount to add when calculating log-fold changes.}

\item{method}{Either 'edgeR' or 'limma'. If set to 'limma', voom is used to
transform the counts and estimate observation weights before applying
limma. In this case, the results also contain the standard errors of the
logFCs.}

\item{csawnorm}{Logical scalar. If \code{TRUE}, uses 
\code{csaw::normOffsets()} to estimate offsets for use with \code{edgeR}, 
rather than the default TMM-based size factors.}
}
\description{
Calculate logFCs and associated p-values for a given comparison, using the
Negative Binomial quasi-likelihood framework provided by edgeR. The observed
counts for the WT variant will be used as offsets in the model.
}
\author{
Charlotte Soneson, Michael Stadler
}
