% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseMutantsByAA.R
\name{collapseMutantsByAA}
\alias{collapseMutantsByAA}
\title{Collapse counts by mutated amino acid}
\usage{
collapseMutantsByAA(se, nameCol = "mutantNameAA")
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}
generated by \code{\link{summarizeExperiment}}}

\item{nameCol}{A character scalar providing the column of 
\code{rowData(se)} that contains the amino acid mutant names (that will 
be the new row names).}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} where
    counts have been aggregated by the mutated amino acid(s).
}
\description{
Collapse counts by mutated amino acid
}
\examples{
se <- readRDS(system.file("extdata", "GSE102901_cis_se.rds",
                          package = "mutscan"))[1:200, ]
## The rows of this object correspond to individual codon variants
dim(se)
head(rownames(se))
## Collapse by amino acid
sec <- collapseMutantsByAA(se)
## The rows of the collapsed object correspond to amino acid variants
dim(sec)
head(rownames(sec))
## The mutantName column contains the individual codon variants that were 
## collapsed
head(rowData(sec))

}
\author{
Charlotte Soneson, Michael Stadler
}
