% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseMutantsByAA.R
\name{collapseMutantsByAA}
\alias{collapseMutantsByAA}
\title{Collapse counts by mutated amino acid}
\usage{
collapseMutantsByAA(se, collapseSynonymous = FALSE)
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}
generated by \code{\link{summarizeExperiment}}}

\item{collapseSynonymous}{A logical scalar indicating whether to represent
all mutations that are synonymous with the wildtype as unmutated}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} where
  counts have been aggregated by the mutated amino acid(s).
}
\description{
Collapse counts by mutated amino acid
}
\author{
Charlotte Soneson, Michael Stadler
}
