% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestFastqs.R
\name{digestFastqs}
\alias{digestFastqs}
\title{Read, filter and digest sequences from two fastq files.}
\usage{
digestFastqs(experimentType, fastqForward, fastqReverse, skipForward = 1,
  skipReverse = 1, umiLengthForward = 10, umiLengthReverse = 8,
  constantLengthForward = 18, constantLengthReverse = 20,
  variableLengthForward = 96, variableLengthReverse = 96,
  adapterForward = "", adapterReverse = "", wildTypeForward = "",
  wildTypeReverse = "", constantForward = "", constantReverse = "",
  avePhredMin = 20, variableNMax = 0, umiNMax = 0,
  nbrMutatedCodonsMax = 1, forbiddenMutatedCodons = "NNW",
  mutatedPhredMin = 0, verbose = FALSE)
}
\arguments{
\item{experimentType}{character(1), either "cis" or "trans". If this is set
to "cis", the variable sequences from the forward and reverse reads will be
consolidated into one single sequence.}

\item{fastqForward, fastqReverse}{character(1), paths to FASTQ files
corresponding to forward and reverse reads, respectively.}

\item{skipForward, skipReverse}{numeric(1), the number of bases to skip in the
start of each forward and reverse read, respectively.}

\item{umiLengthForward, umiLengthReverse}{numeric(1), the length of the
barcode (UMI) sequence in the forward/reverse reads, respectively, not
including the skipped bases (defined by
\code{skipForward}/\code{skipReverse}).}

\item{constantLengthForward, constantLengthReverse}{numeric(1), the length of
the constant sequence in the forward/reverse reads, respectively.}

\item{variableLengthForward, variableLengthReverse}{numeric(1), the length of
the variable sequence in the forward/reverse reads, respectively.}

\item{adapterForward, adapterReverse}{character(1), the adapter sequence for
forward/reverse reads, respectively. If a forward/reverse read contains the
corresponding adapter sequence, the sequence pair will be filtered out.
If set to \code{NULL}, no adapter filtering is performed. The number of
filtered read pairs are reported in the return value.}

\item{wildTypeForward, wildTypeReverse}{character(1), the wild type sequence
for the forward and reverse variable region.}

\item{constantForward, constantReverse}{character(1), the expected constant
forward and reverse sequences.}

\item{avePhredMin}{numeric(1) Minimum average Phred score in the variable
region for a read to be retained. If L contains both forward and reverse
variable regions, the minimum average Phred score has to be achieved in
both for a read pair to be retained.}

\item{variableNMax}{numeric(1) Maximum number of Ns allowed in the variable
region for a read to be retained.}

\item{umiNMax}{numeric(1) Maximum number of Ns allowed in the UMI for a read
to be retained.}

\item{nbrMutatedCodonsMax}{numeric(1) Maximum number of mutated codons that
are allowed.}

\item{forbiddenMutatedCodons}{character vector. Codons (can contain ambiguous
IUPAC characters, see \code{\link[Biostrings]{IUPAC_CODE_MAP}}). If a read
pair contains a mutated codon matching this pattern, it will be filtered
out.}

\item{mutatedPhredMin}{numeric(1) Minimum Phred score of a mutated base for the
read to be retained. If any mutated base has a Phred score lower than
\code{mutatedPhredMin}, the read will be discarded.}

\item{verbose}{logical(1), whether to print out progress messages.}
}
\value{
A list with ---more details---.
}
\description{
\code{readAndDigestFastqs} reads sequences for a pair of fastq files
and digests them (extracts umis, constant and variable parts, filters,
extracts mismatch information from constant and counts the observed
unique variable parts).
}
\details{
more details.
}
