% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{findMismatchPositions}
\alias{findMismatchPositions}
\title{Find positions of mismatching bases in DNAStringSet}
\usage{
findMismatchPositions(pattern, subject, maxChunkSize = 1e+05)
}
\arguments{
\item{pattern}{A nucleotide sequence corresponding to the 
wildtype sequence}

\item{subject}{A DNAStringSet with observed variable sequences}

\item{maxChunkSize}{numeric(1), largest allowed chunk size for steps where
the reads are processed in chunks.}
}
\value{
A two-element list with base/codon positions of mismatches. The
  list elements have classes InterList and NumericList, respectively.
}
\description{
Compare sequences in a DNAStringSet to expected sequence
and return a list with positions of mismatches.
}
\keyword{internal}
