% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterReads.R
\name{mismatchFilter}
\alias{mismatchFilter}
\title{Filter reads based on mismatches with the wildtype sequence}
\usage{
mismatchFilter(wildTypeSeq, variableSeq, nbrMutatedCodonsMax = NULL,
  forbiddenMutatedCodon = NULL)
}
\arguments{
\item{wildTypeSeq}{character(1), the wild type sequence}

\item{variableSeq}{qualityScaledDNAStringSet, containing the sequences of the
variable region}

\item{nbrMutatedCodonsMax}{numeric(1) Maximum number of mutated codons that
are allowed.}

\item{forbiddenMutatedCodon}{character(1) A codon (can contain ambiguous
IUPAC characters, see \code{\link[Biostrings]{IUPAC_CODE_MAP}}). If a read
pair contains a mutated codon matching this pattern, it will be filtered
out.}
}
\value{
A list with two elements:
  \describe{
    \item{minQualMutated}{Lowest quality for the mutated bases}
    \item{keepMM}{Logical vector, whether to keep each read}
 }
}
\description{
Filter reads based on mismatches with the wildtype sequence
}
\author{
Charlotte Soneson
}
\keyword{internal}
