% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Construct a volcano plot}
\usage{
plotVolcano(
  res,
  padjThreshold = 0.05,
  pointSize = "small",
  interactivePlot = FALSE
)
}
\arguments{
\item{res}{\code{data.frame} (typically output from 
\code{calculateRelativeFC()}) with columns corresponding to the 
log-fold change (\code{logFC}), p-value (\code{PValue} or 
\code{P.Value}) and significance (\code{FDR} or \code{adj.P.Val}).}

\item{padjThreshold}{Numeric scalar indicating the adjusted p-value 
threshold to use for coloring the points. All features with 
adjusted p-value below the treshold will be shown in red.}

\item{pointSize}{Either \code{"small"} or \code{"large"}, indicating 
which of the two available plot styles that will be used.}

\item{interactivePlot}{Logical scalar, indicating whether an 
interactive plot should be returned, in which one can hover 
over the individual points and obtain further information.}
}
\value{
If \code{interactivePlot} is \code{TRUE}, a \code{plotly} 
    object. If \code{interactivePlot} is \code{FALSE}, a \code{ggplot2}
    object.
}
\description{
Construct a volcano plot
}
\author{
Charlotte Soneson
}
