% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFastqs.R
\name{readFastqs}
\alias{readFastqs}
\title{Read FASTQ files from CIS or TRANS experiment}
\usage{
readFastqs(experimentType, fastqForward, fastqReverse, skipForward = 1,
  skipReverse = 1, umiLengthForward = 10, umiLengthReverse = 8,
  constantLengthForward = 18, constantLengthReverse = 20,
  variableLengthForward = 96, variableLengthReverse = 96,
  adapterForward = NULL, adapterReverse = NULL, verbose = FALSE)
}
\arguments{
\item{experimentType}{Character(1), either "cis" or "trans".}

\item{fastqForward, fastqReverse}{Character(1), FASTQ files corresponding to 
forward and reverse reads, respectively.}

\item{skipForward, skipReverse}{Numeric(1), the number of bases to skip in the
start of each read.}

\item{umiLengthForward, umiLengthReverse}{Numeric(1), the length of the
barcode (UMI) sequence in the forward/reverse reads, respectively, not
including the skipped bases (defined by \code{skipForward}/\code{skipReverse}).}

\item{constantLengthForward, constantLengthReverse}{Numeric(1), the lengths of
the constant sequence in the forward/reverse reads, respectively.}

\item{variableLengthForward, variableLengthReverse}{Numeric(1), the lengths of 
the variable sequence in the forward/reverse reads, respectively.}

\item{adapterForward, adapterReverse}{Character(1), the adapter sequences for
forward/reverse reads, respectively. If a forward/reverse read contains the
corresponding adapter sequence, the sequence pair will be filtered out.
\code{NULL} does not perform any filtering. The number of filtered read
pairs are reported in the return value.}

\item{verbose}{Logical(1), whether to print out progress messages.}
}
\value{
A list with six elements: \describe{ 
  \item{umis}{Merged forward and reverse UMI sequences}
  \item{constantSeqForward}{Constant forward sequence}
  \item{constantSeqReverse}{Constant reverse sequence}
  \item{variableSeqForward}{Variable forward sequence}
  \item{variableSeqReverse}{Variable reverse sequence}
  \item{readSummary}{data.frame tabulating the experiment type, the total
  number of read pairs, and the number of read pairs filtered out because of
  matches to adapter sequence(s)}
}
}
\description{
It is assumed that both the forward and reverse reads are of the form UMI -
constant sequence - variable sequence, and that the length of the UMI
sequence, and the length of the constant sequence, is the same across all the
reads in the same file (but it can be different for the forward and reverse 
reads).
}
\author{
Charlotte Soneson
}
