% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeExperiment.R
\name{summarizeExperiment}
\alias{summarizeExperiment}
\title{Summarize and collapse multiple mutational scanning experiments}
\usage{
summarizeExperiment(x, meta)
}
\arguments{
\item{x}{A named list of "sequences" objects as returned by \code{\link{readFastqs}}
or \code{\link{filterReads}}). Names are used to link the objects to the
metadata provided in \code{meta}.}

\item{meta}{A \code{data.frame} with at least two columns ("Name" and "Condition").
"Name" will be used to link to objects in \code{x}, and a potentially subset
and reordered version of \code{meta} is stored in \code{colData} or the
returned \code{\link[SummarizedExperiment]{SummarizedExperiment}}.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with
  \describe{
    \item{assays(x)$counts}{containing the observed number of sequences or sequence pairs.}
    \item{rowData(x)}{containing the unique sequences or sequence pairs.}
    \item{colData(x)}{containing the metadata provided by \code{meta}.}
  }
}
\description{
Combine multiple sequence lists (as returned by \code{\link{readFastqs}} or
\code{\link{filterReads}}) into a \code{\link[SummarizedExperiment]{SummarizedExperiment}},
with observed variable sequences (sequence pairs) in rows and samples in
columns.
}
\author{
Michael Stadler
}
